requirements.txt
astroid==3.1.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.8
distlib==0.3.8
filelock==3.13.3
identify==2.5.35
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mysql-connector-python==8.3.0
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
pylint==3.1.0
pytest==8.1.1
PyYAML==6.0.1
requests==2.31.0
requests-mock==1.12.1
setuptools==69.2.0
soupsieve==2.5
tomlkit==0.12.4
urllib3==2.2.1
virtualenv==20.25.1
